+++ server/bin/www
@@ -31,9 +31,12 @@ server.on("error", onError);
 server.on("listening", onListening);
 server.on("upgrade", onUpgrade);
 <<<<<<< HEAD
+<<<<<<< HEAD
 =======
 // startSocketServer(server);
 >>>>>>> 4ee58dd (ws scratch work)
+=======
+>>>>>>> 1be80b7 (Remove cruft)
 
 /**
  * Normalize a port into a number, string, or false.
